EZP-24304: Provide a way to customize field definition form views
> https://jira.ez.no/browse/EZP-24304

Added `getFieldDefinitionEditConfig()` to FieldTypeFormMapperInterface.
Returned value is a hash which may contain:
* `template`: template where specific form fields defined in mapFieldDefinitionForm() will be displayed.
* `vars`: Hash of additional variables to pass to the template.

Provided template will be included during FieldDefinition rendering, in
order to add potential additional fields.
Variables passed to the form are:
* `data`: FieldDefinitionData object
* `languageCode`: Language code as passed to the main form
* `contentTypeDraft`: The ContentTypeDraft object
* `form`: Current FieldDefinition form

Note that it is *not mandatory* to provide a template. If a FieldType
doesn't need additional form fields (i.e. no validator or field
settings), there is no need to define a template.

Also moved update_content_type.html.twig to ContentType subdirectory