Fix EZP-24383: Implement form action dispatcher
> https://jira.ez.no/browse/EZP-24383

This PR introduces a simple mechanism to handle several actions in a single form.
Every time a form using the mechanism is submitted (and validated), a `FormActionEvent`
is fired. This event contains the form, its underlying data, the clicked button name and the language code.

In the case of ContentType, event names are `contentType.update.<buttonName>`, which makes it very easy
to attach a specific *form processor* (which is actually an event subscriber).

`FormDataMapperInterface` is also introduced to help converting repository value objects to the underlying form data.
This will help avoiding doing this in the controller.